Split face_detector into preprocessing/landmarks/mask; drop dead pick_nasion and get_facial_region_mask_from_nasion
