+++ src/cedalion/geometry/photogrammetry/anonymization/__init__.py
@@ -7,58 +7,55 @@ scientific utility.
 
 The pipeline works as follows:
 
-1. User clicks the nasion (Nz) via ``pick_nasion``
-2. Normalize axes so Y=anterior, Z=left
-3. Isolate the head (remove body/shoulders)
-4. Detect remaining landmarks (Iz, Cz, LPA, RPA) from nasion
-5. Align the frame from the 5 landmarks and detect the cap boundary
-6. Build the face + ear deletion mask from the landmarks
-7. Delete the masked vertices
+1. User picks the 5 landmarks (Nz, Iz, Cz, LPA, RPA) via the upstream
+   ``cedalion.plots.plot_surface(pick_landmarks=True)`` picker.
+2. Normalize axes so Y=anterior, Z=left (``preprocessing``).
+3. Isolate the head and remove disconnected fragments (``preprocessing``).
+4. Align the full frame from the 5 landmarks (``preprocessing``).
+5. Detect the cap boundary, build the deletion mask, and delete the masked
+   vertices (``mask``).
 
-An alternative MediaPipe-based automatic pipeline lives on the
-``auto-detection-pipeline`` branch.
+``detect_landmarks_from_nasion`` (``landmarks``) is an alternative that takes
+just the nasion and geometrically infers the other four. An alternative
+MediaPipe-based automatic pipeline lives on the ``auto-detection-pipeline``
+branch.
 
 Example:
     >>> from cedalion.geometry.photogrammetry.anonymization import (
-    ...     pick_nasion, normalize_axes, isolate_head,
-    ...     detect_landmarks_from_nasion, align_axes_from_landmarks,
+    ...     normalize_axes, isolate_head, align_axes_from_landmarks,
     ...     detect_cap_boundary, face_mask_from_landmarks,
     ...     delete_masked_vertices,
     ... )
-    >>> nasion = pick_nasion(surface)
-    >>> surface, nasion, R = normalize_axes(surface, nasion)
-    >>> surface, _ = isolate_head(surface, nasion)
-    >>> landmarks = detect_landmarks_from_nasion(surface, nasion)
+    >>> surface, nz, R = normalize_axes(surface, landmarks.sel(label='Nz'))
+    >>> surface, _ = isolate_head(surface, nz)
+    >>> surface, landmarks, R = align_axes_from_landmarks(surface, landmarks)
 
 Initial Contributors:
     - Face Anonymization Project | 2024
 """
 
-from .face_detector import (
+from .preprocessing import (
     normalize_axes,
     isolate_head,
-    detect_landmarks_from_nasion,
-    get_facial_region_mask_from_nasion,
     align_axes_from_landmarks,
+)
+from .landmarks import detect_landmarks_from_nasion
+from .mask import (
     detect_cap_boundary,
     face_mask_from_landmarks,
     delete_masked_vertices,
 )
-from .ui import pick_nasion
 from .validator import validate_anonymization
 
 
 __all__ = [
-    # Nasion detection
-    "pick_nasion",
-    # Axis normalization and head isolation
+    # Preprocessing (axis normalization, head isolation, full alignment)
     "normalize_axes",
     "isolate_head",
-    # Landmark and face mask detection
-    "detect_landmarks_from_nasion",
-    "get_facial_region_mask_from_nasion",
-    # Landmark-only geometric pipeline
     "align_axes_from_landmarks",
+    # Landmark detection
+    "detect_landmarks_from_nasion",
+    # Mask construction and application
     "detect_cap_boundary",
     "face_mask_from_landmarks",
     "delete_masked_vertices",
